# openapi.yaml
openapi: 3.0.0
info:
  title: 'Cube-3 API'
  contact:
    email: leo.paillard@gmail.com
  version: 1.0.0
servers:
  -
    url: 'http://cyber-dodo.fr:8080/api/v1'
    description: 'Cube-3 API'
paths:
...
  /company-settings:
    post:
      tags:
        - 'Company Settings'
      summary: 'Add company settings'
      operationId: 94126c8d3844338dfdc29cb8671270ee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanySettingsRequest'
      responses:
        '201':
          description: 'Company settings created'
        '400':
          description: 'Invalid request data'
        '404':
          description: 'Company not found'
        '409':
          description: 'Company settings already exist'
        '500':
          description: 'Company settings could not be created'
  '/company-settings/{id}':
    get:
      tags:
        - 'Company Settings'
      summary: 'Get company settings by id'
      operationId: 2aa66e75d711bc80b3eba526111660a5
      parameters:
        -
          name: id
          in: path
          description: 'ID of the company settings'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Company settings found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanySettingsResponse'
        '404':
          description: 'Company settings not found'
        '500':
          description: 'Internal server error'
    put:
      tags:
        - 'Company Settings'
      summary: 'Update company settings by id'
      operationId: 0105213f6dcebc4348c3e38981bd8bfc
      parameters:
        -
          name: id
          in: path
          description: 'ID of the company settings'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'Company settings object that needs to be updated'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanySettingsRequest'
      responses:
        '200':
          description: 'Company settings updated'
        '400':
          description: 'Invalid request data'
        '404':
          description: 'Company settings not found'
        '409':
          description: 'Company settings already exist'
        '500':
          description: 'Company settings could not be created'
    delete:
      tags:
        - 'Company Settings'
      summary: 'Delete company settings by id'
      operationId: 3733f4e95b12a9b46b849dc2f9683746
      parameters:
        -
          name: id
          in: path
          description: 'ID of the company settings'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Company settings deleted'
        '404':
          description: 'Company settings not found'
        '500':
          description: 'Internal server error'
...
components:
  schemas:
...
    CompanyResponse:
      properties:
        id:
          type: integer
          example: '1'
        name:
          type: string
          example: 'Cube 3'
        address:
          type: string
          example: '1, rue de la paix'
        city:
          type: string
          example: Paris
        country:
          type: string
          example: France
        zipCode:
          type: string
          example: '75000'
        phone:
          type: string
          example: '0123456789'
        slogan:
          type: string
          example: 'The best company ever'
        logoPath:
          type: string
          example: cube3.png
        license:
          $ref: '#/components/schemas/LicenseResponse'
        language:
          type: string
          example: fr
        licenseExpirationDate:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        createdAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
    CompanySettingsRequest:
      required:
        - primaryColor
        - secondaryColor
        - tertiaryColor
        - company
      properties:
        primaryColor:
          type: string
          example: '#000000'
        secondaryColor:
          type: string
          example: '#000000'
        tertiaryColor:
          type: string
          example: '#000000'
        company:
          type: id
          example: 1
      type: object
    CompanySettingsResponse:
      properties:
        id:
          type: integer
        primaryColor:
          type: string
          example: '#000000'
        secondaryColor:
          type: string
          example: '#000000'
        tertiaryColor:
          type: string
          example: '#000000'
        company:
          $ref: '#/components/schemas/CompanyResponse'
        created_at:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updated_at:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...
    LicenseResponse:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: basic
        description:
          type: string
          example: 'basic license for 1 user'
        price:
          type: integer
          example: 100
        maxUsers:
          type: integer
          example: 1
        validityPeriod:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
          example: '2021-09-30 12:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-09-30 12:00:00'
      type: object
...
  securitySchemes:
    bearer:
      type: http
      bearerFormat: JWT